+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 
 - :star2: Split chain execution into2 public function for more flexibility.
 - :star2: Each execution has dedicated folder with it's logs.
+- :star2: You can download all files used during an execution from the interface.
 - :wrench: Deprecation on console commands.
-
 
 # 1.0.0 Alpha #1
 - :confetti_ball: :tada: First release :tada: :confetti_ball: